Fix time remaining broken by message pruning
Currently, the time remaining display breaks because the referenced system message gets pruned. This change grabs the ending time needed to calculate the time remaining on load and keeps it for all future references when updating the display.

It also displays "Ending Soon" instead of "0 minutes remaining" when the system message does not indicate a specific number of minutes, but rather 'Voting will end soon'.

Lastly, if the end time was not successfully captured AND we're not 'ending soon', then we hide the time remaining div.