Show version number if GM_info supported
Previously we removed the version string from the settings box in case the user's script engine didn't support it. This instead checks for its availability, and if available, uses it to append the version number.

Also, in the unlikely chance the maintainer wants to increment the version number in two places, they are free to pre-fill the versionNumber variable with the current version number. Since doing so is less than ideal, I'd recommend only bothering to do so on major versions changes, e.g.: 1.8xx -> 1.9xx.