Re-color user name list occasionally
When a user goes/comes back from idle they lose their color because their style attribute is lost when updated. My change sets the coloring on a timer (at an interval to rarely intersect with the `update()` one) and modifies the selector to only attempt to style usernames that are not already styled.