Fix time left not showing
In some cases (maybe RES + Firefox, that's what I use) the time left is not showing up.
It happens because when the script is run the system chat message isn't there yet and it is never checked again afterwards. With this edit the time left is checked again on update when it hasn't been found before.