catchup to main
* Re-color user name list occasionally

When a user goes/comes back from idle they lose their color because their style attribute is lost when updated. My change sets the coloring on a timer (at an interval to rarely intersect with the `update()` one) and modifies the selector to only attempt to style usernames that are not already styled.

* Fix hiding time left display when 'ending soon'

Forgot to check is we're ending soon before hiding the time remaining display box.

* Switch user list name re-coloring to use MutationObserver

Changed from the timed intervals to an observer.

* Styling support for RES Night Mode

Styling to work with RES's Night Mode, including some styling for things this script specifically does.

Only gets inserted if the user has RES installed, and only does any styling if they have Night Mode enabled or enable it. Will disable if they toggle Night Mode off, as all styles are scoped to Night Mode specifically.

Inline via JS isn't my preferred way to do styling, but I think this is the best way to get it to the most users quickly and have it "just work" the easiest, especially while Robin is even still a thing.

* bump version

* Fix spelling mistake

* added ability to unmute users from settings

* custom spam filters