Styling consolidated and general cleanup
Moved all the CSS that didn't need to be inline to a `GM_addStyle` block below the JS to keep the actual code clean and make changing styling much less of a headache. Don't mind the `\` escaping the EOL chars. Did some spacing and lines cleanup. Also added Consolas to the fonts list since the others look terrible on Windows, and it'll just get skipped when not on Windows.